requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# AI/ML libraries
cohere>=4.40
langchain>=0.1.0
langchain-cohere>=0.0.5
google-generativeai>=0.8.0

# Music/Audio processing
spotipy>=2.22.1
librosa>=0.10.0
soundfile>=0.12.1

# Data processing
requests>=2.31.0
python-dotenv>=1.0.0

# Lyrics search and web scraping
lyricsgenius>=3.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
googlesearch-python>=1.1.0

# JSON handling (built-in, but listing for clarity)
json5>=0.9.0

# Date/Time handling (built-in datetime module used)
python-dateutil>=2.8.0

# File handling
pathlib>=1.0.0

# Web scraping (if needed for music data)
googlesearch-python>=1.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: Database support
sqlite3>=0.0.0  # Built-in

# Optional: Enhanced JSON handling
ujson>=5.8.0

# Development dependencies (optional)
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0